backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
bcrypt==4.3.0
    # via -r requirements.in
click==8.2.1
    # via uvicorn
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.116.1
    # via -r requirements.in
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
jinja2==3.1.6
    # via sqladmin
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   wtforms
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic[email]==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via -r requirements.in
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via sqladmin
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqladmin==0.21.0
    # via -r requirements.in
sqlalchemy==2.0.42
    # via
    #   -r requirements.in
    #   alembic
    #   sqladmin
starlette==0.47.2
    # via
    #   fastapi
    #   sqladmin
typing-extensions==4.14.1
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via -r requirements.in
wtforms==3.1.2
    # via sqladmin
